Divided script transpiling into dev/build
This commit adds a `build-dev` task as default in Gulp to allow code-based inspections of type. For example, one might want to log `this.constructor.name` with will produce a non-useful result when the code is uglified. The dev build will not uglify the code to allow for these types of inspections.